LEARNING MANAGEMENT SYSTEM/code_executor_service/requirements.txt
Django>=5.2.9
djangorestframework>=3.16.1